Moved business logic to a custom hook usePoolsSummary + fix lint issues
